+++ packages/plugin-hint/src/index.ts
@@ -13,5 +13,8 @@ export const hintPlugin: StatefulPlugin<typeof hintState> = {
   state: hintState,
   title: 'Hinweis',
   description: 'Gib zusätzliche Tipps zur Aufgabe in dieser ausklappbaren Box.',
-  icon: createIcon(faLightbulb)
+  icon: createIcon(faLightbulb),
+  getFocusableChildren(state) {
+    return [state().content]
+  }
 }