+++ packages/plugin-solution/src/index.ts
@@ -14,5 +14,8 @@ export const solutionPlugin: StatefulPlugin<typeof solutionState> = {
   icon: createIcon(faCheckSquare),
   title: 'Lösung',
   description:
-    'Gestalte in dieser ausklappbaren Box eine ausführliche Lösung zu deinen Aufgaben.'
+    'Gestalte in dieser ausklappbaren Box eine ausführliche Lösung zu deinen Aufgaben.',
+  getFocusableChildren(state) {
+    return [state().content]
+  }
 }